feat(experimental-utils): extract `ast-utils`' `predicates`' helpers
